️🛵📩 ↝ [SSG-190 SSG-191 SSG-193]: Users can now upgrade their telescope FROM the telescope modal
